requirements.txt
pyyaml
htmlmin
